algorithm updated

+++ evaluator-service/src/main/java/com/stackroute/plasma/domain/Evaluator.java
@@ -34,7 +34,7 @@ public class Evaluator {
     @JsonProperty("level")
     private String level;
     @JsonProperty("confidenceScore")
-    private Double confidenceScore;
+    private Float confidenceScore;
 
 
     @Override
